Extract dot to a different module to avoid circular dependencies
