+++ yabt/graph.py
@@ -25,9 +25,8 @@ from collections import defaultdict
 from os.path import relpath
 
 import networkx
-from networkx.algorithms import dag, descendants
+from networkx.algorithms import dag
 
-from yabt import caching
 from .buildfile_parser import process_build_file
 from .compat import walk
 from .config import Config
@@ -38,12 +37,6 @@ from .utils import yprint
 
 
 logger = make_logger(__name__)
-
-
-TARGETS_COLORS = {'Python': 'red', 'PythonTest': 'pink',
-                  'PythonPackage': 'purple', 'CppLib': 'blue',
-                  'AptPackage': 'brown4', 'CostomInstaller': 'brown',
-                  'Proto': 'green'}
 
 
 def stable_reverse_topological_sort(graph):
@@ -328,32 +321,6 @@ def populate_targets_graph(build_context, conf: Config):
     logger.info('Finished parsing build graph with {} nodes and {} edges',
                 build_context.target_graph.order(),
                 build_context.target_graph.size())
-
-
-def write_dot(build_context, conf: Config, out_f):
-    """Write build graph in dot format to `out_f` file-like object."""
-    buildenvs = set(
-        target.buildenv for target in build_context.targets.values()
-        if target.buildenv is not None)
-    buildenv_targets = set(buildenvs)
-    for buildenv in buildenvs:
-        buildenv_targets = buildenv_targets.union(
-            descendants(build_context.target_graph, buildenv))
-    out_f.write('strict digraph  {\n')
-    for node in build_context.target_graph.nodes:
-        if conf.show_buildenv_deps or node not in buildenv_targets:
-            cached = caching.load_target_from_cache(
-                build_context.targets[node], build_context)[0]
-            fillcolor = 'fillcolor="grey",style=filled' if cached else ''
-            color = TARGETS_COLORS.get(
-                build_context.targets[node].builder_name, 'black')
-            out_f.write('  "{}" [color="{}",{}];\n'.format(node, color,
-                                                           fillcolor))
-    out_f.writelines('  "{}" -> "{}";\n'.format(u, v)
-                     for u, v in build_context.target_graph.edges
-                     if conf.show_buildenv_deps or
-                     (u not in buildenv_targets and v not in buildenv_targets))
-    out_f.write('}\n\n')
 
 
 def topological_sort(graph: networkx.DiGraph):
